+++ src/frontend/apps/impress/src/components/quick-search/QuickSearchGroup.tsx
@@ -1,7 +1,7 @@
 import { Command } from 'cmdk';
 import { ReactNode } from 'react';
 
-import { Box } from '../Box';
+import { Box, Text } from '@/components';
 
 import { QuickSearchData } from './QuickSearch';
 import { QuickSearchItem } from './QuickSearchItem';
@@ -23,6 +23,7 @@ export const QuickSearchGroup = <T,>({
         key={group.groupName}
         heading={group.groupName}
         forceMount={false}
+        contentEditable={false}
       >
         {group.startActions?.map((action, index) => {
           return (
@@ -58,7 +59,13 @@ export const QuickSearchGroup = <T,>({
           );
         })}
         {group.emptyString && group.elements.length === 0 && (
-          <span className="ml-b clr-greyscale-500">{group.emptyString}</span>
+          <Text
+            $variation="500"
+            $margin={{ left: '2xs', bottom: '3xs' }}
+            $size="sm"
+          >
+            {group.emptyString}
+          </Text>
         )}
       </Command.Group>
     </Box>