+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteSuggestionMenu.tsx
@@ -43,7 +43,7 @@ export const BlockNoteSuggestionMenu = () => {
     );
     const newSlashMenuItems = [
       ...defaultMenu.slice(0, index + 1),
-      ...getInterlinkingMenuItems(t),
+      ...getInterlinkingMenuItems(editor, t),
       ...defaultMenu.slice(index + 1),
     ];